Added support to fetch and load precompiled Python actor modules
Added a new GET_ACTOR_MODULE command used to request a precomiled
actor module from a peer runtime. When received, the actor module
is written to disc/flash and pending actors of that type are loaded.

Set CC_ACTOR_MODULES_DIR to set the path to store actor modules.

The version of the MicroPython repository has also been updated.